Add api dir, 'reducer'
