+++ client/control/components/DeleteRecipe.js
@@ -1,6 +1,7 @@
 import React, { PropTypes as pt } from 'react';
 import { push } from 'react-router-redux';
-import { makeApiRequest, recipeDeleted } from 'control/actions/ControlActions';
+import makeApiRequest from 'control/api';
+import { recipeDeleted } from 'control/actions/RecipeActions';
 import composeRecipeContainer from 'control/components/RecipeContainer';
 
 class DeleteRecipe extends React.Component {
