+++ client/control/components/RecipeList.js
@@ -4,12 +4,12 @@ import { push } from 'react-router-redux';
 import { Table, Thead, Th, Tr, Td } from 'reactable';
 
 import cloneArrayValues from 'client/utils/clone-array';
+import makeApiRequest from 'control/api';
 
 import {
-  makeApiRequest,
   recipesReceived,
   setSelectedRecipe,
-} from 'control/actions/ControlActions';
+} from 'control/actions/RecipeActions';
 
 import {
   getActiveColumns,
