+++ client/control/reducers/RecipesReducer.js
@@ -6,7 +6,7 @@ import {
   RECIPE_DELETED,
   SET_SELECTED_RECIPE,
   RECIPES_NEED_FETCH,
-} from 'control/actions/ControlActions';
+} from 'control/actions/RecipeActions';
 
 import cloneArrayValues from 'client/utils/clone-array';
